Fix issue with circular import
